New version of Sandbox.
Status: coded, needing last testing.

Now the code can be stored in $HOME and all the large files are automatically copied in $SCRATCH -tested and succesful

The code also now can automatically compress the 6min output and send the copying jobs to the queue with a dependency on the postprocessing job.
The files are compressed with compress_minute_output.py and can be decompressed with decompress_minute_output.py
Needs testing